Fix identification of title element
Previously, `div`s in  METS' logical struct map have been tested
for `DMDID` which in the DSDK data is only present in title elements.
Testing with METS from the SLUB revealed that `DMDID` may well be
present in "real" structure elements. Therefore the test has been
modified to check for `ADMID` which in both collections is realted
to the title element.